@@ -1,41 +1,72 @@
-import React, { Component } from 'react';
-
-import CounterControl from '../../components/CounterControl/CounterControl';
-import CounterOutput from '../../components/CounterOutput/CounterOutput';
+import React, { Component } from "react";
+import { connect } from "react-redux";
+import CounterControl from "../../components/CounterControl/CounterControl";
+import CounterOutput from "../../components/CounterOutput/CounterOutput";
 
 class Counter extends Component {
-    state = {
-        counter: 0
-    }
+  state = {
+    counter: 0
+  };
 
-    counterChangedHandler = ( action, value ) => {
-        switch ( action ) {
-            case 'inc':
-                this.setState( ( prevState ) => { return { counter: prevState.counter + 1 } } )
-                break;
-            case 'dec':
-                this.setState( ( prevState ) => { return { counter: prevState.counter - 1 } } )
-                break;
-            case 'add':
-                this.setState( ( prevState ) => { return { counter: prevState.counter + value } } )
-                break;
-            case 'sub':
-                this.setState( ( prevState ) => { return { counter: prevState.counter - value } } )
-                break;
-        }
+  counterChangedHandler = (action, value) => {
+    switch (action) {
+      case "inc":
+        this.setState(prevState => {
+          return { counter: prevState.counter + 1 };
+        });
+        break;
+      case "dec":
+        this.setState(prevState => {
+          return { counter: prevState.counter - 1 };
+        });
+        break;
+      case "add":
+        this.setState(prevState => {
+          return { counter: prevState.counter + value };
+        });
+        break;
+      case "sub":
+        this.setState(prevState => {
+          return { counter: prevState.counter - value };
+        });
+        break;
     }
+  };
 
-    render () {
-        return (
-            <div>
-                <CounterOutput value={this.state.counter} />
-                <CounterControl label="Increment" clicked={() => this.counterChangedHandler( 'inc' )} />
-                <CounterControl label="Decrement" clicked={() => this.counterChangedHandler( 'dec' )}  />
-                <CounterControl label="Add 5" clicked={() => this.counterChangedHandler( 'add', 5 )}  />
-                <CounterControl label="Subtract 5" clicked={() => this.counterChangedHandler( 'sub', 5 )}  />
-            </div>
-        );
-    }
+  render() {
+    return (
+      <div>
+        <CounterOutput value={this.props.cte} />
+        <CounterControl
+          label="Increment"
+          clicked={this.props.onIncrementCounter}
+        />
+        <CounterControl
+          label="Decrement"
+          clicked={() => this.counterChangedHandler("dec")}
+        />
+        <CounterControl
+          label="Add 5"
+          clicked={() => this.counterChangedHandler("add", 5)}
+        />
+        <CounterControl
+          label="Subtract 5"
+          clicked={() => this.counterChangedHandler("sub", 5)}
+        />
+      </div>
+    );
+  }
 }
 
-export default Counter;
+const mapStateToProps = state => {
+  return {
+    cte: state.counter
+  };
+};
+
+const mapDispatchToProps=dispatch=>{
+  return{
+    onIncrementCounter:()=>dispatch({type:'INCREMENT'})
+  }
+}
+export default connect(mapStateToProps,mapDispatchToProps)(Counter);
